CTDirect API

Index

CTDirect.direct_transcription

CTDirect.set_initial_guess

CTDirect.solve

Documentation

CTDirect.direct_transcription — Method

direct_transcription(
    ocp::CTModels.Model,
    description...;
    grid_size,
    disc_method,
    time_grid,
    init,
    adnlp_backend,
    solver_backend,
    show_time,
    matrix_free
) -> Tuple{CTDirect.DOCP{_A, CTModels.Model{TimesModelType, StateModelType, ControlModelType, VariableModelType, DynamicsModelType, ObjectiveModelType, ConstraintsModelType}} where {_A<:CTDirect.Discretization, TimesModelType<:CTModels.AbstractTimesModel, StateModelType<:CTModels.AbstractStateModel, ControlModelType<:CTModels.AbstractControlModel, VariableModelType<:CTModels.AbstractVariableModel, DynamicsModelType<:Function, ObjectiveModelType<:CTModels.AbstractObjectiveModel, ConstraintsModelType<:CTModels.AbstractConstraintsModel}, ADNLPModels.ADNLPModel{Float64, Vector{Float64}, Vector{Int64}}}

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

Arguments

ocp: optimal control problem as defined in CTModels

[description]: can specifiy for instance the NLP model and / or solver (:ipopt, :madnlp or :knitro)

Keyword arguments (optional)

grid_size: number of time steps for the discretized problem ([250])

disc_method: discretization method ([:trapeze], :euler, :euler_implicit, :midpoint, gauss_legendre_2, gauss_legendre_3)

time_grid: explicit time grid (can be non uniform)

init: info for the starting guess (values as named tuple or existing solution)

adnlp_backend: backend for automatic differentiation in ADNLPModels ([:optimized], :manual, :default)

show_time: (:true, [:false]) show timing details from ADNLPModels

CTDirect.set_initial_guess — Method

set_initial_guess(docp::CTDirect.DOCP, nlp, init) -> Any

Set initial guess in the DOCP

CTDirect.solve — Method

solve(
    ocp::CTModels.Model,
    description::Symbol...;
    display,
    grid_size,
    disc_method,
    time_grid,
    init,
    adnlp_backend,
    kwargs...
) -> CTModels.Solution{TimeGridModelType, TimesModelType, StateModelType, ControlModelType, VariableModelType, CostateModelType, Float64, CTModels.DualModel{CTModels.var"#69#81"{Matrix{Float64}}, CTModels.var"#70#82"{Matrix{Float64}}, Vector{Float64}, Vector{Float64}, CTModels.var"#71#83"{Matrix{Float64}, Int64}, CTModels.var"#72#84"{Matrix{Float64}, Int64}, CTModels.var"#73#85"{Matrix{Float64}, Int64}, CTModels.var"#74#86"{Matrix{Float64}, Int64}, Vector{Float64}, Vector{Float64}}, CTModels.SolverInfos{Dict{Symbol, Any}}} where {TimeGridModelType<:CTModels.TimeGridModel, TimesModelType<:CTModels.TimesModel, StateModelType<:Union{CTModels.StateModelSolution{TS} where TS<:CTModels.var"#63#75", CTModels.StateModelSolution{TS} where TS<:CTModels.var"#64#76"}, ControlModelType<:Union{CTModels.ControlModelSolution{TS} where TS<:CTModels.var"#65#77", CTModels.ControlModelSolution{TS} where TS<:CTModels.var"#66#78"}, VariableModelType<:Union{CTModels.VariableModelSolution{Vector{Float64}}, CTModels.VariableModelSolution{Float64}}, CostateModelType<:Union{CTModels.var"#67#79", CTModels.var"#68#80"}}

Solve an OCP with a direct method

Arguments

ocp: optimal control problem as defined in CTBase

[description]: can specifiy for instance the NLP model and / or solver (:ipopt, :madnlp or :knitro)

Keyword arguments (optional)

display: ([true], false) will disable output if set to false

grid_size: number of time steps for the discretized problem ([250])

disc_method: discretization method ([:trapeze], :midpoint, gauss_legendre_2)

time_grid: explicit time grid (can be non uniform)

init: info for the starting guess (values or existing solution)

adnlp_backend: backend for automatic differentiation in ADNLPModels ([:optimized], :manual, :default)

All further keywords are passed to the inner call of solve_docp