Internal functions

CTDirect.DOCP_constraints! — Method

DOCP_constraints!(c, xu, docp) -> Any

Compute the constraints C for the DOCP problem (modeled as LB <= C(X) <= UB).

CTDirect.DOCP_constraints_check! — Method

DOCP_constraints_check!(cb, constraints, docp)

Check the nonlinear constraints violation for the DOCP problem.

CTDirect.DOCP_initial_guess — Function

DOCP_initial_guess(docp) -> Any
DOCP_initial_guess(docp, init::OptimalControlInit) -> Any

Build initial guess for discretized problem

CTDirect.DOCP_variables_check! — Method

DOCP_variables_check!(vb, variables, docp)

Check the variables box constraints violation for the DOCP problem.

CTDirect.constraints_bounds — Method

constraints_bounds(docp) -> Tuple{Any, Any}

Build upper and lower bounds vectors for the DOCP nonlinear constraints.

CTDirect.get_control_at_time_step — Method

get_control_at_time_step(xu, docp, i) -> Any

Retrieve control variables at given time step from the NLP variables

CTDirect.get_final_time — Method

get_final_time(xu, docp) -> Any

Retrieve final time for OCP (may be fixed or variable)

CTDirect.get_initial_time — Method

get_initial_time(xu, docp) -> Any

Retrieve initial time for OCP (may be fixed or variable)

CTDirect.get_lagrange_cost_at_time_step — Method

get_lagrange_cost_at_time_step(xu, docp, i) -> Any

Retrieve the additional state variable corresponding to the lagrange (running) cost at given time step from the NLP variables

CTDirect.get_state_at_time_step — Method

get_state_at_time_step(xu, docp, i::Int64) -> Any

Retrieve state variables at given time step from the NLP variables

CTDirect.get_time_at_time_step — Method

get_time_at_time_step(xu, docp, i) -> Any

Get actual (un-normalized) time at give time step

CTDirect.get_unnormalized_time — Method

get_unnormalized_time(xu, docp, t_normalized) -> Any

Get actual (un-normalized) time value

CTDirect.get_variable — Method

get_variable(xu, docp) -> Any

Retrieve optimization variables from the NLP variables

CTDirect.parse_DOCP_solution_primal — Method

parse_DOCP_solution_primal(
    docp,
    solution
) -> Tuple{Any, Any, Any}

Recover OCP primal variables from DOCP solution

CTDirect.scalarize — Method

scalarize(arg) -> Any

Convert size 1 vector to scalar for ocp functions if needed

CTDirect.setPathConstraintsAtTimeStep! — Method

setPathConstraintsAtTimeStep!(
    docp,
    index,
    target;
    c,
    args,
    lb,
    ub,
    sol
) -> Any

Set the path constraints / bounds for given time step target = :constraints | :bounds | :multipliers

CTDirect.setPunctualConditions! — Method

setPunctualConditions!(
    docp,
    index,
    target;
    c,
    args_0,
    args_f,
    lb,
    ub
) -> Any

Set the boundary and variable constraints / their bounds target = :constraints | :bounds

CTDirect.setStateEquationAtTimeStep! — Method

setStateEquationAtTimeStep!(
    docp,
    c,
    index,
    args_i,
    args_ip1
) -> Any

Set the constraints corresponding to the state equation

CTDirect.set_control_at_time_step! — Method

set_control_at_time_step!(xu, u_init, docp, i) -> Any

Set control variables at given time step in the NLP variables (for initial guess)

CTDirect.set_state_at_time_step! — Method

set_state_at_time_step!(xu, x_init, docp, i) -> Any

Set state variables at given time step in the NLP variables (for initial guess)

CTDirect.set_variable! — Method

set_variable!(xu, v_init, docp) -> Any

Set optimization variables in the NLP variables (for initial guess)

CTDirect.variables_bounds — Method

variables_bounds(docp) -> Tuple{Any, Any}

Build upper and lower bounds vectors for the DOCP variable box constraints.

CTDirect.vectorize — Method

vectorize(arg) -> Any

Convert scalar from ocp functions to size 1 vector if needed