IRK

Index

CTDirect.Gauss_Legendre_1

CTDirect.Gauss_Legendre_2

CTDirect.Gauss_Legendre_3

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.IRK_dims

CTDirect.runningCost

CTDirect.setStepConstraints!

CTDirect.setWorkArray

Documentation

CTDirect.Gauss_Legendre_1 — Type

Implicit Midpoint discretization, formulated as a generic IRK (ie Gauss Legendre 1) For testing purpose only, use :midpoint instead (cf midpoint.jl) !

CTDirect.Gauss_Legendre_2 — Type

Gauss Legendre 2 discretization, formulated as a generic IRK

CTDirect.Gauss_Legendre_3 — Type

Gauss Legendre 3 discretization, formulated as a generic IRK

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.IRK_dims — Method

IRK_dims(
    dim_NLP_steps,
    dim_NLP_x,
    dim_NLP_u,
    dim_NLP_v,
    dim_path_cons,
    dim_boundary_cons,
    stage
) -> NTuple{5, Any}

Return the dimension of the NLP variables and constraints for a generic IRK discretizion, with the control taken constant per step (ie not distinct controls at time stages)

CTDirect.runningCost — Method

runningCost(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    xu,
    v,
    time_grid
) -> Any

Compute the running cost

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps (+1)

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations