Trapeze

Index

CTDirect.DOCP_Hessian_pattern

CTDirect.DOCP_Jacobian_pattern

CTDirect.runningCost

CTDirect.setStepConstraints!

CTDirect.setWorkArray

Documentation

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{CTDirect.Trapeze}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{CTDirect.Trapeze}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.runningCost — Method

runningCost(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    xu,
    v,
    time_grid
) -> Any

Compute the running cost

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations