Utils

Index

CTDirect.package_loaded

CTDirect.version

Documentation

CTDirect.package_loaded — Method

package_loaded(pkg::String) -> Bool

Check whether a package with the given name is currently loaded.

Arguments

pkg::String: The name of the package to check.

Returns

is_loaded::Bool: true if the package is loaded, false otherwise.

Example

julia> package_loaded("LinearAlgebra")
true

CTDirect.version — Method

version() -> String

Return the version number of the current package as a string.

Returns

version::String: The version of the current module.

Example

julia> version()
"1.2.3"