CTDirect API

Index

CTDirect.OptimalControlInit

CTDirect.available_methods

CTDirect.directTranscription

CTDirect.is_solvable

CTDirect.solve

Available methods

using CTDirect
available_methods()

(:adnlp, :ipopt)

Documentation

CTDirect.OptimalControlInit — Type

Initialization of the optimal control problem solution for the NLP problem.

Constructors:

OptimalControlInit(): default initialization

OptimalControlInit(x_init, u_init, v_init): constant vector or function handles

OptimalControlInit(sol): from existing solution

Examples

julia> init = OptimalControlInit()
julia> init = OptimalControlInit(x_init=[0.1, 0.2], u_init=0.3)
julia> init = OptimalControlInit(x_init=[0.1, 0.2], u_init=0.3, v_init=0.5)
julia> init = OptimalControlInit(x_init=[0.1, 0.2], u_init=t->sin(t), v_init=0.5)
julia> init = OptimalControlInit(sol)

CTDirect.available_methods — Method

available_methods() -> Tuple{Tuple{Vararg{Symbol}}}

Return the list of available methods to solve the optimal control problem.

CTDirect.directTranscription — Method

directTranscription(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size
) -> CTDirect.DOCP

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Return if the optimal control problem ocp is solvable or not by the method solve.

CTDirect.solve — Method

solve(
    docp::CTDirect.DOCP;
    init,
    display,
    print_level,
    mu_strategy,
    kwargs...
) -> OptimalControlSolution

Solve a discretized optimal control problem