CTDirect API

Index

CTDirect.DOCP

CTDirect._OptimalControlInit

CTDirect.OCPSolutionFromDOCP

CTDirect.available_methods

CTDirect.directTranscription

CTDirect.export_OCP_solution

CTDirect.getNLP

CTDirect.is_solvable

CTDirect.load_OCP_solution

CTDirect.read_OCP_solution

CTDirect.save_OCP_solution

CTDirect.setInitialGuess

Available methods

using CTDirect
available_methods()

(:adnlp, :ipopt)

Documentation

CTDirect.DOCP — Type

Struct for discretized optimal control problem DOCP

Contains:

a copy of the original OCP

a NLP formulation of the DOCP

data required to link the two problems

CTDirect._OptimalControlInit — Type

Initial guess for OCP, contains

functions of time for the state and control variables

vector for optimization variables

Initialization data for each field can be left to default or:

vector for optimization variables

constant / vector / function for state and control

existing solution ('warm start') for all fields

Constructors:

_OptimalControlInit(): default initialization

_OptimalControlInit(state, control, variable, time): constant vector, function handles and / or matrices / vectors interpolated along given time grid

_OptimalControlInit(sol): from existing solution

Examples

julia> init = _OptimalControlInit()
julia> init = _OptimalControlInit(state=[0.1, 0.2], control=0.3)
julia> init = _OptimalControlInit(state=[0.1, 0.2], control=0.3, variable=0.5)
julia> init = _OptimalControlInit(state=[0.1, 0.2], controlt=t->sin(t), variable=0.5)
julia> init = _OptimalControlInit(state=[[0, 0], [1, 2], [5, -1]], time=[0, .3, 1.], controlt=t->sin(t))
julia> init = _OptimalControlInit(sol)

CTDirect.OCPSolutionFromDOCP — Method

OCPSolutionFromDOCP(
    docp,
    docp_solution_ipopt
) -> OptimalControlSolution

Build OCP functional solution from DOCP vector solution (given as a GenericExecutionStats)

CTDirect.available_methods — Method

available_methods() -> Tuple{Tuple{Vararg{Symbol}}}

Return the list of available methods to solve the optimal control problem.

CTDirect.directTranscription — Method

directTranscription(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid
) -> DOCP

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

CTDirect.getNLP — Method

getNLP(docp::DOCP) -> Any

Extract the NLP problem from the DOCP

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Check if an OCP is solvable by the method solve.

CTDirect.setInitialGuess — Method

setInitialGuess(docp::DOCP, init) -> Any

Set initial guess in the DOCP

<! – manually add docstrings from package extensions –> <! – does not work for solve which is not in CTDirect –>

CommonSolve.solve — Function

CommonSolve.solve(args...; kwargs...)

Solves an equation or other mathematical problem using the algorithm specified in the arguments. Generally, the interface is:

CommonSolve.solve(prob::ProblemType,alg::SolverType; kwargs...)::SolutionType

where the keyword arguments are uniform across all choices of algorithms.

By default, solve defaults to using solve! on the iterator form, i.e.:

solve(args...; kwargs...) = solve!(init(args...; kwargs...))

solve(
    docp::DOCP;
    init,
    display,
    print_level,
    mu_strategy,
    linear_solver,
    kwargs...
) -> SolverCore.GenericExecutionStats

Solve a discretized optimal control problem DOCP

solve(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid,
    display,
    print_level,
    mu_strategy,
    linear_solver,
    kwargs...
) -> OptimalControlSolution

Solve an optimal control problem OCP by direct method

CTDirect.save_OCP_solution — Function

save_OCP_solution(
    sol::OptimalControlSolution;
    filename_prefix
)

Save OCP solution in JLD2 format

CTDirect.load_OCP_solution — Function

load_OCP_solution() -> Any
load_OCP_solution(filename_prefix) -> Any

Load OCP solution in JLD2 format

CTDirect.export_OCP_solution — Function

export_OCP_solution(
    sol::OptimalControlSolution;
    filename_prefix
)

Export OCP solution in JSON format

CTDirect.read_OCP_solution — Function

read_OCP_solution() -> Any
read_OCP_solution(filename_prefix) -> Any

Read OCP solution in JSON format