CTDirect API

Index

CTDirect.DOCP

CTDirect.available_methods

CTDirect.build_solution

CTDirect.build_solution

CTDirect.direct_transcription

CTDirect.export_ocp_solution

CTDirect.get_nlp

CTDirect.import_ocp_solution

CTDirect.is_solvable

CTDirect.set_initial_guess

CommonSolve.solve

Documentation

CTDirect.DOCP — Type

Struct for discretized optimal control problem DOCP

Contains:

a copy of the original OCP

a NLP formulation of the DOCP

data required to link the two problems

CTDirect.available_methods — Method

available_methods() -> Tuple{Tuple{Vararg{Symbol}}}

Return the list of available methods to solve the optimal control problem.

CTDirect.build_solution — Method

build_solution(
    docp,
    docp_solution_ipopt
) -> OptimalControlSolution

Build OCP functional solution from DOCP discrete solution (given as a GenericExecutionStats)

CTDirect.build_solution — Method

build_solution(docp; primal, dual)

Build OCP functional solution from the DOCP discrete solution, given as a vector. Costate will be retrieved from dual variables (multipliers) if available.

CTDirect.direct_transcription — Method

direct_transcription(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid
) -> DOCP

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

CTDirect.get_nlp — Method

get_nlp(docp::DOCP) -> Any

Extract the NLP problem from the DOCP

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Check if an OCP is solvable by the method solve.

CTDirect.set_initial_guess — Method

set_initial_guess(docp::DOCP, init) -> Any

Set initial guess in the DOCP

CommonSolve.solve — Function

CommonSolve.solve(args...; kwargs...)

Solves an equation or other mathematical problem using the algorithm specified in the arguments. Generally, the interface is:

CommonSolve.solve(prob::ProblemType,alg::SolverType; kwargs...)::SolutionType

where the keyword arguments are uniform across all choices of algorithms.

By default, solve defaults to using solve! on the iterator form, i.e.:

solve(args...; kwargs...) = solve!(init(args...; kwargs...))

solve(
    docp::DOCP;
    init,
    display,
    print_level,
    mu_strategy,
    max_iter,
    tol,
    linear_solver,
    kwargs...
) -> SolverCore.GenericExecutionStats

Solve a discretized optimal control problem DOCP

solve(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid,
    display,
    print_level,
    mu_strategy,
    max_iter,
    tol,
    linear_solver,
    kwargs...
) -> OptimalControlSolution

Solve an optimal control problem OCP by direct method

CTDirect.export_ocp_solution — Function

export_ocp_solution(
    sol::OptimalControlSolution;
    filename_prefix
)

Export OCP solution in JSON format

CTDirect.import_ocp_solution — Function

import_ocp_solution() -> Any
import_ocp_solution(filename_prefix) -> Any

Read OCP solution in JSON format