CTDirect API

Index

CTDirect.available_methods

CTDirect.direct_solve

CTDirect.direct_transcription

CTDirect.export_ocp_solution

CTDirect.import_ocp_solution

CTDirect.is_solvable

CTDirect.set_initial_guess

Documentation

CTDirect.available_methods — Method

available_methods(

) -> Tuple{Tuple{Symbol, Symbol}, Tuple{Symbol, Symbol}}

Return the list of available methods to solve the optimal control problem.

CTDirect.direct_solve — Method

direct_solve(
    ocp::OptimalControlModel,
    description::Symbol...;
    init,
    grid_size,
    time_grid,
    kwargs...
) -> OptimalControlSolution

Solve an OCP with a direct method

CTDirect.direct_transcription — Method

direct_transcription(
    ocp::OptimalControlModel,
    description...;
    init,
    grid_size,
    time_grid
) -> Tuple{CTDirect.DOCP, ADNLPModels.ADNLPModel{Float64, Vector{Float64}, Vector{Int64}}}

Discretize an optimal control problem into a nonlinear optimization problem (ie direct transcription)

CTDirect.is_solvable — Method

is_solvable(ocp) -> Bool

Check if an OCP is solvable by the method solve.

CTDirect.set_initial_guess — Method

set_initial_guess(docp::CTDirect.DOCP, nlp, init) -> Any

Set initial guess in the DOCP

CTDirect.export_ocp_solution — Function

export_ocp_solution(
    sol::OptimalControlSolution;
    filename_prefix
)

Export OCP solution in JSON format

CTDirect.import_ocp_solution — Function

import_ocp_solution(
    ocp::OptimalControlModel;
    filename_prefix
) -> OptimalControlSolution

Read OCP solution in JSON format