Private API

This page lists non-exported (internal) symbols of CTDirect.

From CTDirect

Scheme

CTDirect.Scheme — Type

abstract type Scheme

Abstract type representing a discretization scheme strategy for an optimal control problem.

Concrete subtypes of Scheme define specific schemes for transforming a continuous-time problem into a discrete-time representation suitable for numerical solution.

Example

julia> struct MyScheme <: Scheme end
MyScheme