Internal functions

CTDirect.DOCP — Type

Struct for discretized optimal control problem DOCP

Contains:

a copy of the original OCP

a NLP formulation of the DOCP

data required to link the two problems

CTDirect.DOCP_constraints! — Method

DOCP_constraints!(c, xu, docp) -> Any

Compute the constraints C for the DOCP problem (modeled as LB <= C(X) <= UB).

CTDirect.DOCP_constraints_check! — Method

DOCP_constraints_check!(cb, constraints, docp)

Check the nonlinear constraints violation for the DOCP problem.

CTDirect.DOCP_variables_check! — Method

DOCP_variables_check!(vb, variables, docp)

Check the variables box constraints violation for the DOCP problem.

CTDirect.OCPSolutionFromDOCP_raw — Method

OCPSolutionFromDOCP_raw(
    docp,
    solution;
    objective,
    constraints_violation,
    iterations,
    multipliers_constraints,
    multipliers_LB,
    multipliers_UB,
    message
)

Build OCP functional solution from DOCP vector solution (given as raw variables and multipliers plus some optional infos)

CTDirect.constraints_bounds — Method

constraints_bounds(docp) -> Tuple{Any, Any}

Build upper and lower bounds vectors for the DOCP nonlinear constraints.

CTDirect.get_control_at_time_step — Method

get_control_at_time_step(xu, docp, i) -> Any

Retrieve control variables at given time step from the NLP variables

CTDirect.get_final_time — Method

get_final_time(xu, docp) -> Any

Retrieve final time for OCP (may be fixed or variable)

CTDirect.get_initial_time — Method

get_initial_time(xu, docp) -> Any

Retrieve initial time for OCP (may be fixed or variable)

CTDirect.get_lagrange_cost_at_time_step — Method

get_lagrange_cost_at_time_step(xu, docp, i) -> Any

Retrieve the additional state variable corresponding to the lagrange (running) cost at given time step from the NLP variables

CTDirect.get_state_at_time_step — Method

get_state_at_time_step(xu, docp, i::Int64) -> Any

Retrieve state variables at given time step from the NLP variables

CTDirect.get_variable — Method

get_variable(xu, docp) -> Any

Retrieve optimization variables from the NLP variables

CTDirect.initial_guess — Function

initial_guess(docp) -> Any
initial_guess(docp, init::OptimalControlInit) -> Any

Build initial guess for discretized problem

CTDirect.parse_DOCP_solution — Method

parse_DOCP_solution(
    docp,
    solution,
    multipliers_constraints,
    multipliers_LB,
    multipliers_UB,
    constraints
) -> NTuple{18, Any}

Parse DOCP solution into OCP variables, constraints and multipliers

CTDirect.set_control_at_time_step! — Method

set_control_at_time_step!(xu, u_init, docp, i) -> Any

Set control variables at given time step in the NLP variables (for initial guess)

CTDirect.set_state_at_time_step! — Method

set_state_at_time_step!(xu, x_init, docp, i) -> Any

Set state variables at given time step in the NLP variables (for initial guess)

CTDirect.set_variable! — Method

set_variable!(xu, v_init, docp) -> Any

Set optimization variables in the NLP variables (for initial guess)

CTDirect.variables_bounds — Method

variables_bounds(docp) -> Tuple{Any, Any}

Build upper and lower bounds vectors for the DOCP variable box constraints.