Internal functions

CTBase.OptimalControlSolution — Method

OptimalControlSolution(
    docp::CTDirect.DOCP,
    docp_solution
) -> OptimalControlSolution

Build OCP functional solution from DOCP discrete solution (given as a SolverCore.GenericExecutionStats)

CTBase.OptimalControlSolution — Method

OptimalControlSolution(
    docp::CTDirect.DOCP;
    primal,
    dual,
    objective,
    iterations,
    constraints_violation,
    message,
    mult_LB,
    mult_UB
) -> OptimalControlSolution

Build OCP functional solution from the DOCP discrete solution, given as a vector. Costate will be retrieved from dual variables (multipliers) if available.

CTDirect.DOCP — Type

Struct for discretized optimal control problem DOCP

Contains:

a copy of the original OCP

data required to link the OCP with the discretized DOCP

CTDirect.Euler_explicit — Type

Explicit Euler discretization, formulated as a generic IRK For testing purpose only, use :euler instead (cf euler.jl) !

CTDirect.Euler_implicit — Type

Implicit Euler discretization, formulated as a generic IRK For testing purpose only, use :euler_implicit instead (cf euler.jl) !

CTDirect.Gauss_Legendre_1 — Type

Implicit Midpoint discretization, formulated as a generic IRK (ie Gauss Legendre 1) For testing purpose only, use :midpoint instead (cf midpoint.jl) !

CTDirect.Gauss_Legendre_2 — Type

Gauss Legendre 2 discretization, formulated as a generic IRK

CTDirect.Gauss_Legendre_3 — Type

Gauss Legendre 3 discretization, formulated as a generic IRK

CTDirect.Trapeze_implicit — Type

Implicit Trapeze (Crank-Nicolson) discretization, formulated as a generic IRK For testing purpose only, use :trapeze instead (cf trapeze.jl) !

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{CTDirect.Midpoint}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{CTDirect.Trapeze}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Hessian_pattern — Method

DOCP_Hessian_pattern(
    docp::CTDirect.DOCP{D<:CTDirect.Discretization}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Hessian of Lagrangian

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{CTDirect.Midpoint}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{CTDirect.Trapeze}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_Jacobian_pattern — Method

DOCP_Jacobian_pattern(
    docp::CTDirect.DOCP{D<:CTDirect.Discretization}
) -> SparseArrays.SparseMatrixCSC{Bool, Int64}

Build sparsity pattern for Jacobian of constraints

CTDirect.DOCP_constraints! — Method

DOCP_constraints!(c, xu, docp::CTDirect.DOCP) -> Any

Compute the constraints C for the DOCP problem (modeled as LB <= C(X) <= UB).

CTDirect.DOCP_initial_guess — Function

DOCP_initial_guess(docp::CTDirect.DOCP) -> Vector{Float64}
DOCP_initial_guess(
    docp::CTDirect.DOCP,
    init::OptimalControlInit
) -> Vector{Float64}

Build initial guess for discretized problem

CTDirect.DOCP_objective — Method

DOCP_objective(xu, docp::CTDirect.DOCP) -> Any

Compute the objective for the DOCP problem.

CTDirect.IRK_dims — Method

IRK_dims(
    dim_NLP_steps,
    dim_NLP_x,
    dim_NLP_u,
    dim_NLP_v,
    dim_u_cons,
    dim_x_cons,
    dim_xu_cons,
    dim_boundary_cons,
    dim_v_cons,
    stage
) -> NTuple{5, Any}

Return the dimension of the NLP variables and constraints for a generic IRK discretizion, with the control taken constant per step (ie not distinct controls at time stages)

CTDirect.__adnlp_backend — Method

__adnlp_backend() -> Symbol

Used to set the default backend for AD in ADNLPModels. The default value is :optimized.

CTDirect.__control_type — Method

__control_type() -> Symbol

Used to set the default control type for IRK schemes The default value is :constant.

CTDirect.__disc_method — Method

__disc_method() -> Symbol

Used to set the default discretization method. The default value is trapeze.

CTDirect.__display — Method

__display() -> Bool

Used to set the default display toggle. The default value is true.

CTDirect.__grid_size — Method

__grid_size() -> Int64

Used to set the default grid size. The default value is 250.

CTDirect.__ipopt_linear_solver — Method

__ipopt_linear_solver() -> String

Used to set the default value of the linear solver of Ipopt for the direct method. The default value is mumps.

CTDirect.__ipopt_mu_strategy — Method

__ipopt_mu_strategy() -> String

Used to set the default value of the μ strategy of Ipopt for the direct method. The default value is adaptive.

CTDirect.__ipopt_print_level — Method

__ipopt_print_level() -> Int64

Used to set the default value of the print level of Ipopt for the direct method. The default value is 5.

CTDirect.__knitro_print_level — Method

__knitro_print_level() -> Int64

Used to set the default value of the print level of Knitro for the direct method. The default value is 3.

CTDirect.__madnlp_linear_solver — Method

__madnlp_linear_solver() -> String

Used to set the default value of the linear solver of MadNLP for the direct method. The default value is umfpack.

CTDirect.__max_iterations — Method

__max_iterations() -> Int64

Used to set the default maximum of iterations. The default value is 1000.

CTDirect.__ocp_init — Method

__ocp_init()

Used to set the default initial guess. The default value is nothing and will correspond to 0.1 for all variables.

CTDirect.__time_grid — Method

__time_grid()

Used to set the default time grid. The default value is nothing.

CTDirect.__tolerance — Method

__tolerance() -> Float64

Used to set the default tolerance. The default value is 1e-6.

CTDirect.add_nonzero_block! — Method

add_nonzero_block!(M, i_start, i_end, j_start, j_end; sym)

Add block of nonzeros elements to a sparsity pattern Format: boolean matrix (M) or index vectors (Is, Js) Includes a more compact method for single element case Option to add the symmetric block also (eg for Hessian) Note: independent from discretization scheme

CTDirect.build_bounds — Method

build_bounds(
    dim_var,
    dim_box,
    box_triplet
) -> Tuple{Any, Any}

Build full, ordered sets of bounds for state, control or optimization variables

CTDirect.constraints_bounds! — Method

constraints_bounds!(
    docp::CTDirect.DOCP
) -> Tuple{Vector{Float64}, Vector{Float64}}

Build upper and lower bounds vectors for the DOCP nonlinear constraints.

CTDirect.get_OCP_control_at_time_step — Method

get_OCP_control_at_time_step(
    xu,
    docp::CTDirect.DOCP{<:CTDirect.Discretization, <:CTDirect.ScalVect, CTDirect.ScalVariable},
    i
) -> Any

Retrieve control variables at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps, with convention u(tf) = U_N Scalar / Vector output

CTDirect.get_OCP_control_at_time_step — Method

get_OCP_control_at_time_step(
    xu,
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK, <:CTDirect.ScalVect, CTDirect.ScalVariable},
    i
) -> Any

Retrieve control variables at given time step (/stage) from the NLP variables. Convention: 1 <= i <= dimNLPsteps Scalar / Vector output Step / Stage versions

CTDirect.get_OCP_control_at_time_step — Method

get_OCP_control_at_time_step(
    xu,
    docp::CTDirect.DOCP{CTDirect.Euler, <:CTDirect.ScalVect, CTDirect.ScalVariable},
    i
) -> Any

Retrieve control variables at given time step from the NLP variables. Convention: see above for acplicit / implicit versions Scalar / Vector output

CTDirect.get_OCP_control_at_time_step — Method

get_OCP_control_at_time_step(
    xu,
    docp::CTDirect.DOCP{CTDirect.Trapeze, <:CTDirect.ScalVect, CTDirect.ScalVariable},
    i
) -> Any

Retrieve control variables at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps+1 Scalar / Vector output

CTDirect.get_OCP_state_at_time_step — Method

get_OCP_state_at_time_step(
    xu,
    docp::CTDirect.DOCP{<:CTDirect.Discretization, CTDirect.ScalVariable},
    i
) -> Any

Retrieve state variables at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps+1 Scalar / Vector output

CTDirect.get_OCP_variable — Method

get_OCP_variable(
    xu,
    docp::CTDirect.DOCP{<:CTDirect.Discretization, <:CTDirect.ScalVect, <:CTDirect.ScalVect, CTDirect.ScalVariable}
) -> Any

Retrieve optimization variables from the NLP variables. Convention: stored at the end, hence not dependent on the discretization method Scalar / Vector output

CTDirect.get_lagrange_state_at_time_step — Method

get_lagrange_state_at_time_step(
    xu,
    docp::CTDirect.DOCP,
    i
) -> Any

Retrieve state variable for lagrange cost at given time step from the NLP variables. Convention: 1 <= i <= dimNLPsteps+1

CTDirect.get_stagevars_at_time_step — Method

get_stagevars_at_time_step(
    xu,
    docp::CTDirect.DOCP,
    i,
    j
) -> Any

Retrieve stage variables at given time step/stage from the NLP variables. Convention: 1 <= i <= dimNLPsteps(+1), 1 <= j <= s Vector output Note that passing correct indices is up to the caller, no checks are made here.

CTDirect.parse_DOCP_solution_dual — Method

parse_DOCP_solution_dual(
    docp,
    multipliers,
    constraints
) -> Tuple{Any, Tuple{Any, Any, Any, Vector{Float64}, Vector{Float64}}, Tuple{Any, Any, Any, Vector{Float64}, Vector{Float64}}}

Recover OCP costate and constraints multipliers from DOCP multipliers

CTDirect.parse_DOCP_solution_primal — Method

parse_DOCP_solution_primal(
    docp,
    solution;
    mult_LB,
    mult_UB
) -> Tuple{Any, Any, Any, NTuple{6, Any}}

Recover OCP primal variables from DOCP solution

CTDirect.setPathConstraints! — Method

setPathConstraints!(docp, c, ti, xi, ui, v, offset) -> Any

Set path constraints at given time step

CTDirect.setPathConstraintsBounds! — Method

setPathConstraintsBounds!(
    docp::CTDirect.DOCP,
    lb,
    ub,
    offset
) -> Any

Set bounds for the path constraints at given time step

CTDirect.setPointConstraints! — Method

setPointConstraints!(docp::CTDirect.DOCP, c, xu, v) -> Any

Set the boundary and variable constraints

CTDirect.setPointConstraintsBounds! — Method

setPointConstraintsBounds!(
    docp::CTDirect.DOCP,
    lb,
    ub
) -> Int64

Set bounds for the boundary and variable constraints

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps (+1)

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Euler},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Midpoint},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setStepConstraints! — Method

setStepConstraints!(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    c,
    xu,
    v,
    time_grid,
    i,
    work
) -> Any

Set the constraints corresponding to the state equation Convention: 1 <= i <= dimNLPsteps+1

CTDirect.setWorkArray — Method

setWorkArray(docp::CTDirect.DOCP, xu, time_grid, v) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Euler},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Midpoint},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.setWorkArray — Method

setWorkArray(
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    xu,
    time_grid,
    v
) -> Any

Set work array for all dynamics and lagrange cost evaluations

CTDirect.set_control_at_time_step! — Method

set_control_at_time_step!(
    xu,
    u_init,
    docp::CTDirect.DOCP,
    i
) -> Any

Set initial guess for control variables at given time step Convention: 1 <= i <= dimNLPsteps

CTDirect.set_control_at_time_step! — Method

set_control_at_time_step!(
    xu,
    u_init,
    docp::CTDirect.DOCP{<:CTDirect.GenericIRK},
    i
) -> Any

Set initial guess for control variables at given time step (/stage) Convention: 1 <= i <= dimNLPsteps+1 Step / stage versions

CTDirect.set_control_at_time_step! — Method

set_control_at_time_step!(
    xu,
    u_init,
    docp::CTDirect.DOCP{CTDirect.Trapeze},
    i
) -> Any

Set initial guess for control variables at given time step Convention: 1 <= i <= dimNLPsteps+1

CTDirect.set_optim_variable! — Method

set_optim_variable!(xu, v_init, docp) -> Any

Set optimization variables in the NLP variables (for initial guess)

CTDirect.set_state_at_time_step! — Method

set_state_at_time_step!(
    xu,
    x_init,
    docp::CTDirect.DOCP,
    i
) -> Any

Set initial guess for state variables at given time step Convention: 1 <= i <= dimNLPsteps+1

CTDirect.variables_bounds! — Method

variables_bounds!(
    docp::CTDirect.DOCP
) -> Tuple{Vector{Float64}, Vector{Float64}}

Build upper and lower bounds vectors for the DOCP variable box constraints.